# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-commerce backend. Money is always in minor units: a price_cents of 1999 means $19.99.",
        "title": "go-ecommerce API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Public. Returns a page of active products, newest first, with their category and images.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "List products",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number, from 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size, up to 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A page of products",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductListEnvelope"
                        }
                    },
                    "400": {
                        "description": "Invalid page or limit",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only. The product is created active. Prices are in minor units: 1999 means $19.99.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a product",
                "parameters": [
                    {
                        "description": "Product details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductEnvelope"
                        }
                    },
                    "400": {
                        "description": "Validation failed, or the category does not exist",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Not an administrator",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "The SKU is already in use",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Public. An inactive product is reported as not found.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductEnvelope"
                        }
                    },
                    "400": {
                        "description": "The id is not a number",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "No such product, or it is not on sale",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only. The SKU cannot be changed. Omitting is_active leaves it unchanged.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated product",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductEnvelope"
                        }
                    },
                    "400": {
                        "description": "Validation failed, or the category does not exist",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Not an administrator",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "No such product",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only. The product is soft-deleted and its SKU becomes available again.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.MessageEnvelope"
                        }
                    },
                    "400": {
                        "description": "The id is not a number",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Not an administrator",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "No such product",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CreateProductRequest": {
            "type": "object",
            "required": [
                "category_id",
                "name",
                "price_cents",
                "sku"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string",
                    "maxLength": 5000
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "price_cents": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string",
                    "maxLength": 100
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "email is not available"
                },
                "message": {
                    "type": "string",
                    "example": "invalid credentials"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.MessageEnvelope": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "logged out"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.PageMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductResponse"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductImageResponse": {
            "type": "object",
            "properties": {
                "alt_text": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductListEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.PageMeta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Pointer and omitempty: an unloaded relation is absent rather than an\nobject full of zero values that a client cannot tell from real data.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryResponse"
                        }
                    ]
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ProductImageResponse"
                    }
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price_cents": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.UpdateProductRequest": {
            "type": "object",
            "required": [
                "category_id",
                "name",
                "price_cents"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string",
                    "maxLength": 5000
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "price_cents": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token from /auth/login, sent as \"Bearer \u003ctoken\u003e\". A refresh token is rejected here.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}